ABI_V0 compatibility: graphics.library function offsets and protos

+++ rom/graphics/bltpattern.c
@@ -62,10 +62,10 @@ static ULONG bltpattern_render(APTR bpr_data, WORD srcx, WORD srcy,
 /*  SYNOPSIS */
 	AROS_LHA(struct RastPort *, rp, A1),
 	AROS_LHA(PLANEPTR         , mask, A0),
-	AROS_LHA(WORD             , xMin, D0),
-	AROS_LHA(WORD             , yMin, D1),
-	AROS_LHA(WORD             , xMax, D2),
-	AROS_LHA(WORD             , yMax, D3),
+	AROS_LHA(LONG             , xMin, D0),
+	AROS_LHA(LONG             , yMin, D1),
+	AROS_LHA(LONG             , xMax, D2),
+	AROS_LHA(LONG             , yMax, D3),
 	AROS_LHA(ULONG            , byteCnt, D4),
 
 /*  LOCATION */
